+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageHTMLFull.h
@@ -1,0 +1,8 @@
+#import <Foundation/Foundation.h>
+#import "ABKInAppMessageHTML.h"
+
+/*
+ * Appboy Public API: ABKInAppMessageHTMLFull
+ */
+@interface ABKInAppMessageHTMLFull : ABKInAppMessageHTML
+@end