+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageHTMLFullViewController.h
@@ -1,0 +1,4 @@
+#import "ABKInAppMessageHTMLViewController.h"
+
+@interface ABKInAppMessageHTMLFullViewController : ABKInAppMessageHTMLViewController
+@end