+++ Example/Stopwatch/InAppMessageUIViewController.h
@@ -11,6 +11,9 @@
 @property NSMutableDictionary *inAppMessageDictionary;
 @property IBOutlet UISegmentedControl *inAppMessageTypeSegment;
 @property IBOutlet UITableView *tableView;
+@property IBOutlet UIView *HTMLComposerView;
+@property IBOutlet UITextField *zipRemoteURLTextField;
+@property IBOutlet UITextView *HTMLInAppTextView;
 
 - (IBAction) inAppMessageTypeChanged:(id)sender;